drone actions no longer affect total actions for the group this means that drone actions have no impact against the min/max number of actions for a group refactored roster download code to allow building of non-web versions of the application
