+++ lib/screens/roster/download/download.dart
@@ -1,0 +1,5 @@
+import 'package:gearforce/models/roster/roster.dart';
+
+void downloadRoster(UnitRoster roster) {
+  print('roster downloading not implemented yet on this platform');
+}
